mqtt.py
from paho.mqtt import client as mqtt_client
import time
import random

MQTT_BROKER = ""
MQTT_PORT = 1883

# generate client ID with pub prefix randomly
client_id = f'python-mqtt-{random.randint(0, 1000)}'
username = ""
password = ""

# topic
topic = "test"

def connect_mqtt(broker=MQTT_BROKER, port=MQTT_PORT, client_id=client_id, username=username, password=password):
    def on_connect(client, userdata, flags, rc):
        if rc == 0:
            print("Connected to MQTT Mosquitto Broker!")
        else:
            print("Failed to connect, return code %d", rc)
    
    def on_disconnect(client, userdata, rc):
        if rc != 0:
            print("Unexpected disconnection.")

    # Set Connecting Client ID
    client = mqtt_client.Client(client_id)
    client.username_pw_set(username, password)
    
    return client

def publish(client):
    msg_count = 0
    while True:
        time.sleep(1)
        msg = f"messages: {msg_count}"
        result = client.publish(topic, msg)
        # result: [0, 1]
        status = result[0]
        if status == 0:
            print(f"Send `{msg}` to topic `{topic}`")
        else:
            print(f"Failed to send message to topic {topic}")
        msg_count += 1