requirements.txt
paho_mqtt==1.6.1
SQLAlchemy==1.4.44
psycopg2-binary==2.9.9
